Tigger formula evaluation on setCellFormula call
I've found that when using the `setCellFormula()` function with a formula that references cells with _another_ formula (e.g. calculating the total of a column comprised of price*qty cells), the formula is not evaluated. You can click in the cell and hit enter to get the calculation to work, but to evaluate before writing the spreadsheet to disk, you need to run the `formulaEvaluator`. The getCellValue() function already does this, so I've added a call to that within `setCellFormula()` and this resolves the problem for me.